Refactor functionality of extract_ir.py into a library
This commit refactors most of the functional elements of extract_ir.py
into a library. This serves a couple of purposes. It makes some elements
of the code base a little bit cleaner (no more hacky flag tricks with
absl to get things working in unit tests) and can be seen as making some
things slightly more readable. In addition to this, it enables use of
these functions in other areas, particularly downstream projects within
the compiler_opt python package.
